Rename test file
It’s not ES6 ( https://github.com/stoeffel/chickencurry/issues/11 )
