+++ test.js
@@ -1,2 +1,2 @@
 import './test/es5';
-import './test/es6';
+import './test/es-2016';
